% Generated by build script
\name{fdc_list}
\alias{fdc_list}
\title{List foods by data type}
\usage{
fdc_list()
}
\arguments{
\item{data_type}{One of "Foundation", "SR Legacy", "Survey (FNDDS)",}
\item{page_size}{Results per page (default 25, max 200)}
\item{api_key}{API key (default DEMO_KEY)}
}
\value{tibble: fdc_id, description, data_type, brand_owner}
\description{List foods by data type}
